Rename Optional.h to KSOptional.h
